add index_exists?
Checking for index existence is a common operation, and previously it
was pretty cumbersome.

Instead of getting the status of all indices and checking for the
requested key, this implementation checks for the status of the
requested index only and rescues the missing exception. This saves on
bandwidth if you have a lot of indices.